pyproject.toml
[project]
name = "chester"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv",
  "httpx",
  "python-chess",
  "tqdm",
  "google-genai",
  "modal",
]

[project.scripts]
chester = "chester.cli:main"
